requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
asyncpg
alembic
email-validator
python-jose[cryptography]
bcrypt
python-multipart
passlib
cloudinary
sphinx
shibuya